+++ modules/cover-project/cover-project.php
@@ -1,4 +1,7 @@
-<?php function telabotanica_module_cover_project($data) {
+<?php
+require_once 'inc/filters.php';
+
+function telabotanica_module_cover_project($data) {
 
   echo '<div class="cover-project">';
 
@@ -11,7 +14,9 @@
       $cover_image_url
     );
 
-    bp_group_avatar( 'type=full&width=90&height=90' );
+    if ( ! bp_disable_group_avatar_uploads() ) :
+      bp_group_avatar( 'type=full&width=90&height=90' );
+    endif;
 
     echo '<h1 class="cover-project-title">' . bp_get_group_name() . '</h1>';
     echo '<div class="cover-project-description">' . bp_get_group_description_excerpt() . '</div>';
@@ -39,12 +44,7 @@
         bp_get_group_type(),
         get_telabotanica_module('icon', ['icon' => $visibility_icons[ bp_get_group_status() ] ])
       );
-      the_telabotanica_module('button', [
-        'href' => '#',
-        'text' => __( 'Suivre ce projet', 'telabotanica' ) . get_telabotanica_module('icon', ['icon' => 'plus']),
-        'modifiers' => 'outline'
-      ] );
-      // bp_group_join_button();
+      bp_group_join_button();
       // var_dump( bp_get_group_join_button() );
     echo '</div>';